+++ internal/transport/connector_rpc.go
@@ -48,6 +48,40 @@ func (s *ConnectorRPCServer) Execute(args *ConnectorExecuteArgs, reply *Connecto
 	return nil
 }
 
+func (s *ConnectorRPCServer) OpenStream(args *ConnectorOpenStreamArgs, reply *ConnectorOpenStreamReply) error {
+	resp, err := s.Impl.OpenStream(args.ProviderHandle, args.Request)
+	reply.Response = resp
+	reply.Error = core.EncodePluginError(err)
+	return nil
+}
+
+func (s *ConnectorRPCServer) ReceiveStream(args *ConnectorReceiveStreamArgs, reply *ConnectorReceiveStreamReply) error {
+	resp, err := s.Impl.ReceiveStream(args.ProviderHandle, args.Request)
+	reply.Response = resp
+	reply.Error = core.EncodePluginError(err)
+	return nil
+}
+
+func (s *ConnectorRPCServer) AckStream(args *ConnectorAckStreamArgs, reply *ConnectorAckStreamReply) error {
+	reply.Error = core.EncodePluginError(s.Impl.AckStream(args.ProviderHandle, args.Request))
+	return nil
+}
+
+func (s *ConnectorRPCServer) NackStream(args *ConnectorNackStreamArgs, reply *ConnectorNackStreamReply) error {
+	reply.Error = core.EncodePluginError(s.Impl.NackStream(args.ProviderHandle, args.Request))
+	return nil
+}
+
+func (s *ConnectorRPCServer) GrantCredits(args *ConnectorGrantCreditsArgs, reply *ConnectorGrantCreditsReply) error {
+	reply.Error = core.EncodePluginError(s.Impl.GrantCredits(args.ProviderHandle, args.Request))
+	return nil
+}
+
+func (s *ConnectorRPCServer) CloseStream(args *ConnectorCloseStreamArgs, reply *ConnectorCloseStreamReply) error {
+	reply.Error = core.EncodePluginError(s.Impl.CloseStream(args.ProviderHandle, args.Request))
+	return nil
+}
+
 func (s *ConnectorRPCServer) Probe(args *ConnectorProbeArgs, reply *ConnectorProbeReply) error {
 	event, err := s.Impl.Probe(args.ProviderHandle, args.Request)
 	reply.Event = event
@@ -98,6 +132,58 @@ func (c *ConnectorRPCClient) ExecuteWithTimeout(providerHandle string, req core.
 		return core.ConnectorResponse{}, err
 	}
 	return reply.Response, core.DecodePluginError(reply.Error)
+}
+
+func (c *ConnectorRPCClient) OpenStream(providerHandle string, req core.StreamOpenRequest) (core.StreamOpenResponse, error) {
+	var reply ConnectorOpenStreamReply
+	if err := callWithTimeout(c.Client, "Plugin.OpenStream", &ConnectorOpenStreamArgs{ProviderHandle: providerHandle, Request: req}, &reply, defaultRPCTimeout); err != nil {
+		return core.StreamOpenResponse{}, err
+	}
+	return reply.Response, core.DecodePluginError(reply.Error)
+}
+
+func (c *ConnectorRPCClient) ReceiveStream(providerHandle string, req core.StreamReceiveRequest) (core.StreamReceiveResponse, error) {
+	timeout := defaultRPCTimeout
+	if req.WaitTimeoutMS > 0 {
+		timeout += time.Duration(req.WaitTimeoutMS) * time.Millisecond
+	}
+	var reply ConnectorReceiveStreamReply
+	if err := callWithTimeout(c.Client, "Plugin.ReceiveStream", &ConnectorReceiveStreamArgs{ProviderHandle: providerHandle, Request: req}, &reply, timeout); err != nil {
+		return core.StreamReceiveResponse{}, err
+	}
+	return reply.Response, core.DecodePluginError(reply.Error)
+}
+
+func (c *ConnectorRPCClient) AckStream(providerHandle string, req core.StreamAckRequest) error {
+	var reply ConnectorAckStreamReply
+	if err := callWithTimeout(c.Client, "Plugin.AckStream", &ConnectorAckStreamArgs{ProviderHandle: providerHandle, Request: req}, &reply, defaultRPCTimeout); err != nil {
+		return err
+	}
+	return core.DecodePluginError(reply.Error)
+}
+
+func (c *ConnectorRPCClient) NackStream(providerHandle string, req core.StreamNackRequest) error {
+	var reply ConnectorNackStreamReply
+	if err := callWithTimeout(c.Client, "Plugin.NackStream", &ConnectorNackStreamArgs{ProviderHandle: providerHandle, Request: req}, &reply, defaultRPCTimeout); err != nil {
+		return err
+	}
+	return core.DecodePluginError(reply.Error)
+}
+
+func (c *ConnectorRPCClient) GrantCredits(providerHandle string, req core.StreamGrantCreditsRequest) error {
+	var reply ConnectorGrantCreditsReply
+	if err := callWithTimeout(c.Client, "Plugin.GrantCredits", &ConnectorGrantCreditsArgs{ProviderHandle: providerHandle, Request: req}, &reply, defaultRPCTimeout); err != nil {
+		return err
+	}
+	return core.DecodePluginError(reply.Error)
+}
+
+func (c *ConnectorRPCClient) CloseStream(providerHandle string, req core.StreamCloseRequest) error {
+	var reply ConnectorCloseStreamReply
+	if err := callWithTimeout(c.Client, "Plugin.CloseStream", &ConnectorCloseStreamArgs{ProviderHandle: providerHandle, Request: req}, &reply, defaultRPCTimeout); err != nil {
+		return err
+	}
+	return core.DecodePluginError(reply.Error)
 }
 
 func (c *ConnectorRPCClient) Probe(providerHandle string, req core.ConnectorRequest) (core.ResourceStatusEvent, error) {